pyproject.toml
[tool.poetry]
name = "measurement-pred"
version = "0.1.0"
description = ""
authors = ["Oliver Daniels-Koch <danielskoch.oliver@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.41.0"
datasets = "^2.19.1"
evaluate = "^0.4.2"
accelerate = "^0.30.1"
torch = "^2.3.0"
torchmetrics = "^1.4.0.post0"
numpy = "^1.26.4"
tensorboard = "^2.16.2"
submitit = "^1.5.1"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
hydra-submitit-launcher = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
